+++ xenon-common/src/main/java/com/vmware/xenon/common/OperationContext.java
@@ -65,16 +65,6 @@ public class OperationContext {
         return threadTransactionId.get();
     }
 
-    /**
-     * Sets current thread's authorization context based on {@code op} headers and/or cookies.
-     *
-     * @param host Service host.
-     * @param op   Operation containing authorization headers / cookies.
-     */
-    public static void setAuthorizationContext(ServiceHost host, Operation op) {
-        setAuthorizationContext(host.getAuthorizationContext(op));
-    }
-
     public static AuthorizationContext getAuthorizationContext() {
         return threadAuthContext.get();
     }